requirements.txt
Flask==1.0.2
pyjnius==1.1.1
spacy==2.0.11
